feat: apply safety margin to route recommendations
- Add safetyMarginPercent to recommendation requests with a default of 15 percent
- Use useful flight time for candidate generation, filtering, scoring, warnings, and explanations
- Allow users to edit the safety margin in the frontend form
- Show useful flight time in recommendation metrics
- Improve map rendering for selected routes, waypoints, and departure airport bounds
- Add tests for safety margin filtering and default behavior

+++ src/main/java/flightdiscovery/paull/application/recommendation/RouteCandidateGenerator.java
@@ -33,7 +33,7 @@ public class RouteCandidateGenerator {
 
     public List<FlightRoute> generate(
             Airport departureAirport,
-            int availableTimeMinutes,
+            double availableTimeMinutes,
             double cruiseSpeedKmh,
             String preference
     ) {
@@ -99,7 +99,7 @@ public class RouteCandidateGenerator {
         );
     }
 
-    private boolean fitsAvailableTime(FlightRoute route, int availableTimeMinutes, double cruiseSpeedKmh) {
+    private boolean fitsAvailableTime(FlightRoute route, double availableTimeMinutes, double cruiseSpeedKmh) {
         double distanceKm = routeCalculationService.totalDistanceKm(route);
         double estimatedTimeHours = routeCalculationService.estimatedTimeHours(distanceKm, cruiseSpeedKmh);
         double estimatedTimeMinutes = routeCalculationService.estimatedTimeMinutes(estimatedTimeHours);
